+++ PoporUI/Classes/UINavigationController/UINavigationController+pSize.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface UINavigationController (pSize)
 
-@property (nonatomic) int barHeight;// 隐藏导航栏的视图
+@property (nonatomic) NSInteger barHeight;// 隐藏导航栏的视图
 
 @property (nonatomic) BOOL _initBarHeight; // 内部使用
