+++ PoporUI/Classes/UITextView/UITextView+pMaxLength.h
@@ -30,6 +30,6 @@
 //    }];
 //}
 
-- (void)textViewMaxLength:(int)maxLength block:(void(^)(BOOL isEditing, BOOL isOutRange))textFieldBlock;
+- (void)textViewMaxLength:(NSInteger)maxLength block:(void(^)(BOOL isEditing, BOOL isOutRange))textFieldBlock;
 
 @end
