+++ PoporUI/Classes/UITextView/UITextView+pMaxLength.m
@@ -20,14 +20,14 @@
     [[NSNotificationCenter defaultCenter] removeObserver:target name:UITextViewTextDidChangeNotification object:nil];
 }
 
-- (void)textViewMaxLength:(int)maxLength block:(void(^)(BOOL isEditing, BOOL isOutRange))textFieldBlock
+- (void)textViewMaxLength:(NSInteger)maxLength block:(void(^)(BOOL isEditing, BOOL isOutRange))textFieldBlock
 {
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.01 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [self delayML:maxLength block:textFieldBlock];
     });
 }
 
-- (void)delayML:(int)maxLength block:(void(^)(BOOL isEditing, BOOL isOutRange))textFieldBlock
+- (void)delayML:(NSInteger)maxLength block:(void(^)(BOOL isEditing, BOOL isOutRange))textFieldBlock
 {
     NSString *toBeString = self.text;
